[ADD] pos_customer_wallet: Tests
